requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.25.2
pandas==2.1.1
spacy==3.7.2
neo4j==5.14.0
python-multipart==0.0.6
weaviate-client==3.25.2
pinecone-client==2.2.4
scikit-learn==1.3.2
pytest==7.4.3
httpx==0.25.1
openai==1.2.4
anthropic==0.5.1
matplotlib==3.8.1
networkx==3.2.1
pymupdf==1.23.5  # For PDF parsing
beautifulsoup4==4.12.2  # For HTML parsing
tqdm==4.66.1 